cross fixed

+++ backend/controllers/workoutcontrollers.js
@@ -3,25 +3,20 @@ const mongoose = require("mongoose");
 
 //get all workouts
 const getWorkouts = async (req, res) => {
-  try {
-    const workouts = await Workout.find().sort({ createdAt: -1 });
-    res.status(200).json({ workouts });
-  } catch (err) {
-    res.status(400).json({ err: err.message });
-  }
+  const workouts = await Workout.find().sort({ createdAt: -1 });
+
+  res.status(200).json(workouts);
 };
 
 //get single workout
 const getSingleWorkout = async (req, res) => {
   const { id } = req.params;
 
-  try {
-    const workout = await Workout.findById(id);
-    res.status(200).json({ workout });
-  } catch (err) {
-    res.status(400).json({ err: "workout not found" });
+  const workout = await Workout.findById(id);
+  if (!workout) {
+    return res.status(404).json({ err: "Workout not found" });
   }
-  res.json(getSingleWorkout);
+  res.status(200).json(workout); // Return workout directly
 };
 
 //post new workout
@@ -29,37 +24,39 @@ const postWorkout = async (req, res) => {
   const { title, reps, load } = req.body;
   try {
     const workout = await Workout.create({ title, reps, load });
-    res.status(200).json({ workout });
+    res.status(200).json(workout); // Return workout directly
   } catch (err) {
-    res.status(400).json({ err:'workout not created' });
+    res.status(400).json({ err: "Workout not created" });
   }
 };
 
 //delete workout
 const deleteWorkout = async (req, res) => {
   const { id } = req.params;
-  try {
-    const workout = await Workout.findByIdAndDelete(id);
-    res.status(200).json({ workout });
-  } catch (err) {
-    res.status(400).json({ err: 'workout not found' });
-  }
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).json("No workout with that id  ");
+
+  const workout = await Workout.findByIdAndDelete(id);
+  if (!workout) return res.status(404).json({ err: "Workout not found" });
+
+  res.status(200).json(workout); // Return workout directly
 };
 
 //update workout
 const updateWorkout = async (req, res) => {
   const { id } = req.params;
-  const { title, reps, load } = req.body;
-  try {
-    const workout = await Workout.findByIdAndUpdate(
-      id,
-      { title, reps, load },
-      { new: true }
-    );
-    res.status(200).json({ workout });
-  } catch (err) {
-    res.status(400).json({ err: 'workout not found' });
-  }
+
+  if (!mongoose.Types.ObjectId.isValid(id))
+    return res.status(404).json("No workout with that id  ");
+
+  const workout = await Workout.findByIdAndUpdate(
+    id,
+    { ...req.body },
+    { new: true }
+  );
+  if (!workout) return res.status(404).json({ err: "Workout not found" });
+
+  res.status(200).json(workout); // Return workout directly
 };
 
 module.exports = {
